refact: moving just-spider file to reservation components folder
