+++ frontend/src/app/reservation/CreateReservation/index.tsx
@@ -1,5 +1,5 @@
 import React, { useState, ForwardedRef } from 'react';
-import { JustSpider } from "../../../shared/components/just-spider";
+import { JustSpider } from "../components/just-spider";
 import {
   Box,
   Text,
