+++ frontend/src/app/reservation/SelectReservation/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { JustSpider } from '../../../shared/components/just-spider';
+import { JustSpider } from '../components/just-spider';
 import { Box, Text, Icon, HStack, Button, VStack } from '@chakra-ui/react';
 import { FaArrowLeft, FaWifi, FaCar, FaCoffee, FaSnowflake, FaConciergeBell, FaCheck, FaHeart, FaShareAlt } from 'react-icons/fa';
 import { NavBar } from '../../../shared/components/nav-bar';
